requirements.txt
streamlit == 1.41.1
psycopg2 == 2.9.10
firebase-admin == 6.6.0
python-dotenv == 1.0.1
polars == 1.21.0